Added some more button bindings, Fixed the movecontainerarmtopositionondashboard commmand so it actually does what its named.  This is useful for tuning.  Thats untested, but i'm pretty sure it'll work.

+++ UPS2015/src/org/usfirst/frc/team1675/robot/subsystems/ToteStacker.java
@@ -69,7 +69,7 @@ public class ToteStacker extends PIDSubsystem {
      * 
      * @param distance tells how far the tote stacker moves from it's current point
      */
-    public void moveRelative(int distance){
+    public void moveRelative(double distance){
     	this.setSetpoint(this.getSetpoint() + distance);
     	System.out.println("moveRelative new setpoint: " + this.getSetpoint());
     }
